Update zsh config

+++ uninstall.sh
@@ -1,15 +1,19 @@
 #!/bin/bash
 
-dotfiles=(".tmux.conf" ".zshrc" ".conkyrc" ".bashrc" ".Xresources" ".gitconfig" ".vimrc")
+dotfiles=(".tmux.conf" ".zshrc" ".conkyrc" ".bashrc" ".Xresources" ".gitconfig" ".vimrc" ".zshenv")
 
 for file in "${dotfiles[@]}"; do
-	echo "Removing ${HOME}/${file}"
-	rm $HOME/$file
+	if [ -f "$HOME/$file" ]; then
+		echo "Removing ${HOME}/${file}"
+		rm $HOME/$file
+	fi
 done
 
 directories=(".mpd" ".ncmpcpp" "bin" ".zsh")
 
 for directory in "${directories[@]}"; do
-	echo "Removing ${HOME}/${directory}"
-	rm -r $HOME/$directory
+	if [ -d "$HOME/$directory" ]; then
+		echo "Removing ${HOME}/${directory}"
+		rm -r $HOME/$directory
+	fi
 done
